refactor: extract text-area logic to reusable TextAreaMixin
